requirements.txt
python-dotenv
supervisely
